Add tests for utility functions

+++ util/rand.go
@@ -4,11 +4,11 @@ import "crypto/rand"
 
 const alphabet = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789"
 
-func RandomString(length int) string {
+func RandomString(length uint) string {
 	b := make([]byte, length)
 	rand.Read(b)
 
-	for i := 0; i < length; i++ {
+	for i := uint(0); i < length; i++ {
 		b[i] = alphabet[b[i]%byte(len(alphabet))]
 	}
